# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para registro e consulta de contratos de financiamento automotivo utilizando blockchain Hyperledger Besu e contratos inteligentes UUPS",
        "title": "MobX API - VFinance Registry",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MobX API Support",
            "email": "support@mobx-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "3.0.0"
    },
    "host": "144.22.179.183:3000",
    "basePath": "/",
    "paths": {
...
        "/api/blockchain/admin/brand": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra uma nova marca de veículo (apenas admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write-operations"
                ],
                "summary": "Register Brand",
                "parameters": [
                    {
                        "description": "Dados da marca",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca registrada com sucesso",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/blockchain/admin/config": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza configurações do servidor (apenas admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write-operations"
                ],
                "summary": "Update Server Config",
                "parameters": [
                    {
                        "description": "Configurações do servidor",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuração atualizada com sucesso",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/blockchain/admin/model": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registra um novo modelo de veículo (apenas admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write-operations"
                ],
                "summary": "Register Model",
                "parameters": [
                    {
                        "description": "Dados do modelo",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modelo registrado com sucesso",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/blockchain/brand/{brandId}": {
            "get": {
                "description": "Obtém nome da marca por ID on-chain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blockchain"
                ],
                "summary": "Get Brand Name",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da marca",
                        "name": "brandId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nome da marca",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Brand ID inválido",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Marca não encontrada",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/blockchain/contract/exists/{registryId}": {
            "get": {
                "description": "Verifica se contrato existe on-chain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blockchain"
                ],
                "summary": "Check Contract Exists",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registry ID do contrato",
                        "name": "registryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status de existência",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Registry ID inválido",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/blockchain/contract/metadata": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza o hash de metadados de um contrato existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "write-operations"
                ],
                "summary": "Update Metadata Hash",
                "parameters": [
                    {
                        "description": "Dados para atualização",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hash atualizado com sucesso",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/blockchain/contract/registry/{registryId}": {
            "get": {
                "description": "Busca contrato on-chain por Registry ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blockchain"
                ],
                "summary": "Get Contract by Registry ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registry ID do contrato",
                        "name": "registryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato encontrado",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Registry ID inválido",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Contrato não encontrado",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e o token JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}